requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
jinja2>=3.1.4
python-multipart>=0.0.12
itsdangerous>=2.2.0
pydantic-settings>=2.6.0
